📝 Fix majority of documentation test failures (131→43 failures)
Major progress on doc test fixes:
- Fixed imports in all major modules (lib.rs, neo_builder, neo_clients, etc.)
- Replaced prelude imports with specific module imports
- Fixed incorrect crate name references (NeoRust → neo3)
- Updated API calls to match current implementation
- Fixed type imports and trait requirements

Modules completed:
✅ src/lib.rs - Fixed all main library examples
✅ neo_builder/script/script_builder.rs - All 15 tests passing
✅ neo_builder/transaction/* - signers, transaction_builder, witness_rule
✅ neo_clients/* - RPC client and transport examples
✅ neo_codec/* - binary encoder/decoder examples
✅ neo_crypto/* - key management and base58 examples
✅ neo_protocol/* - account and NEP2 examples
✅ neo_types/* - address, scripthash, and utility examples
✅ neo_wallets/* - wallet management examples (24 tests)
✅ neo_contract, neo_fs, neo_x, neo_utils modules

Remaining: 43 failing tests to fix

+++ src/neo_builder/script/script_builder.rs
@@ -52,7 +52,7 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::ScriptBuilder;
+	/// use neo3::neo_builder::ScriptBuilder;
 	///
 	/// let builder = ScriptBuilder::new();
 	/// ```
@@ -73,7 +73,8 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::{ScriptBuilder, OpCode};
+	/// use neo3::neo_builder::ScriptBuilder;
+	/// use neo3::neo_types::OpCode;
 	///
 	/// let mut builder = ScriptBuilder::new();
 	/// builder.op_code(&[OpCode::Push1, OpCode::Push2, OpCode::Add]);
@@ -99,7 +100,8 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::{ScriptBuilder, OpCode};
+	/// use neo3::neo_builder::ScriptBuilder;
+	/// use neo3::neo_types::OpCode;
 	///
 	/// let mut builder = ScriptBuilder::new();
 	/// builder.op_code_with_arg(OpCode::PushData1, vec![0x01, 0x02, 0x03]);
@@ -127,7 +129,10 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::{ScriptBuilder, H160, ContractParameter, CallFlags};
+	/// use neo3::neo_builder::ScriptBuilder;
+	/// use primitive_types::H160;
+	/// use neo3::neo_types::ContractParameter;
+	/// use neo3::builder::CallFlags;
 	///
 	/// let mut builder = ScriptBuilder::new();
 	/// let contract_hash = H160::from_slice(&[0; 20]);
@@ -174,7 +179,8 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::{ScriptBuilder, InteropService};
+	/// use neo3::neo_builder::ScriptBuilder;
+	/// use neo3::builder::InteropService;
 	///
 	/// let mut builder = ScriptBuilder::new();
 	/// builder.sys_call(InteropService::SystemRuntimeCheckWitness);
@@ -205,7 +211,8 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::{ScriptBuilder, ContractParameter};
+	/// use neo3::neo_builder::ScriptBuilder;
+	/// use neo3::neo_types::ContractParameter;
 	///
 	/// let mut builder = ScriptBuilder::new();
 	/// builder.push_params(&[
@@ -238,7 +245,8 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::{ScriptBuilder, ContractParameter};
+	/// use neo3::neo_builder::ScriptBuilder;
+	/// use neo3::neo_types::ContractParameter;
 	///
 	/// let mut builder = ScriptBuilder::new();
 	/// builder.push_param(&ContractParameter::from("Hello, Neo!")).unwrap();
@@ -300,7 +308,7 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::ScriptBuilder;
+	/// use neo3::neo_builder::ScriptBuilder;
 	/// use num_bigint::BigInt;
 	///
 	/// let mut builder = ScriptBuilder::new();
@@ -376,7 +384,8 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::{ScriptBuilder, OpCode};
+	/// use neo3::neo_builder::ScriptBuilder;
+	/// use neo3::neo_types::OpCode;
 	///
 	/// let mut builder = ScriptBuilder::new();
 	/// builder.push_opcode_bytes(OpCode::PushData1, vec![0x01, 0x02, 0x03]);
@@ -412,7 +421,7 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::ScriptBuilder;
+	/// use neo3::neo_builder::ScriptBuilder;
 	///
 	/// let mut builder = ScriptBuilder::new();
 	/// builder.push_data("Hello, Neo!".as_bytes().to_vec());
@@ -451,7 +460,7 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::ScriptBuilder;
+	/// use neo3::neo_builder::ScriptBuilder;
 	///
 	/// let mut builder = ScriptBuilder::new();
 	/// builder.push_bool(true);
@@ -583,10 +592,11 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::*;
+	/// use neo3::neo_builder::ScriptBuilder;
+	/// use primitive_types::H160;
 	/// use std::str::FromStr;
 	///
-	/// let sender = ScriptHash::from_str("0xd2a4cff31913016155e38e474a2c06d08be276cf").unwrap();
+	/// let sender = H160::from_str("0xd2a4cff31913016155e38e474a2c06d08be276cf").unwrap();
 	/// let nef_checksum = 1234567890;
 	/// let name = "MyContract";
 	///
@@ -633,11 +643,14 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::*;
+	/// use neo3::neo_builder::ScriptBuilder;
+	/// use neo3::neo_types::ContractParameter;
+	/// use neo3::builder::CallFlags;
+	/// use primitive_types::H160;
 	/// use std::str::FromStr;
 	///
 	/// // Call a contract method that returns an iterator and collect up to 100 items
-	/// let contract_hash = ScriptHash::from_str("0xd2a4cff31913016155e38e474a2c06d08be276cf").unwrap();
+	/// let contract_hash = H160::from_str("0xd2a4cff31913016155e38e474a2c06d08be276cf").unwrap();
 	/// let method = "getTokens";
 	/// let params = vec![ContractParameter::from("owner_address")];
 	/// let max_items = 100;
@@ -693,8 +706,9 @@ impl ScriptBuilder {
 		sb.op_code_with_arg(OpCode::JmpIf, vec![0]);
 
 		let jmp_offset = sb.len();
-		let jmp_bytes = (cycle_start - jmp_offset) as u8;
-		sb.op_code_with_arg(OpCode::Jmp, vec![jmp_bytes]);
+		// Calculate backward jump as a signed offset
+		let jmp_bytes = (cycle_start as i32 - jmp_offset as i32) as i8;
+		sb.op_code_with_arg(OpCode::Jmp, vec![jmp_bytes as u8]);
 
 		let load_result = sb.len();
 		sb.op_code(&[OpCode::Nip, OpCode::Nip]);
@@ -721,7 +735,7 @@ impl ScriptBuilder {
 	/// # Examples
 	///
 	/// ```rust
-	/// use neo3::prelude::ScriptBuilder;
+	/// use neo3::neo_builder::ScriptBuilder;
 	///
 	/// let mut builder = ScriptBuilder::new();
 	/// builder.push_data("Hello, Neo!".as_bytes().to_vec());
